+++ _libs/pandora_toolbox/video/src/test_opengl.cpp
@@ -1,5 +1,0 @@
-#include "video/api/opengl.h"
-
-void testOpenGl() {
-  
-}